+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/_help.py
@@ -155,6 +155,19 @@ helps['ad app update'] = """
                         }
                    ]
                 }]
+        - name: declare an application role
+          text: |
+                az ad app update --id e042ec79-34cd-498f-9d9f-123456781234 --app-roles @manifest.json
+                ("manifest.json" contains the following conten)
+                [{
+                    "allowedMemberTypes": [
+                      "User"
+                    ],
+                    "description": "Approvers can mark documents as approved",
+                    "displayName": "Approver",
+                    "isEnabled": "true",
+                    "value": "approver"
+                }]
         - name: update an application's group membership claims to "All"
           text: >
                 az ad app update --id e042ec79-34cd-498f-9d9f-123456781234 --set groupMembershipClaims=All
@@ -340,6 +353,19 @@ helps['ad app create'] = """
                             "type": "Scope"
                         }
                    ]
+                }]
+        - name: Create an application with a role
+          text: |
+                az ad app create --id e042ec79-34cd-498f-9d9f-123456781234 --display-name mytestapp --identifier-uris https://mytestapp.websites.net --app-roles @manifest.json
+                ("manifest.json" contains the following conten)
+                [{
+                    "allowedMemberTypes": [
+                      "User"
+                    ],
+                    "description": "Approvers can mark documents as approved",
+                    "displayName": "Approver",
+                    "isEnabled": "true",
+                    "value": "approver"
                 }]
 """
 helps['ad group'] = """