+++ src/azure-cli-core/HISTORY.rst
@@ -3,6 +3,15 @@
 Release History
 ===============
 
+2.0.60
+++++++
+* Minor fixes.
+
+
+2.0.59
+++++++
+* Fix issue where in some instances using `--subscription NAME` would throw an exception.
+
 2.0.58
 ++++++
 * `az --version` now displays a notification if you have packages that can be updated.